voice_to_image/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

